4th push

+++ ft_memcpy.c
@@ -6,7 +6,7 @@
 /*   By: tmoragli <tmoragli@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/09/06 00:21:17 by telli             #+#    #+#             */
-/*   Updated: 2020/12/07 17:41:43 by tmoragli         ###   ########.fr       */
+/*   Updated: 2020/12/07 18:58:57 by tmoragli         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memcpy(void *dest, const void *src, size_t n)
 {
-	int i;
+	size_t i;
 	unsigned char *cast_dest;
 	unsigned char *cast_src;
